BUGFIX: Ensure account identifier is always added to events
For some events the account identifier is not set since it's
not initialized. Instead of doing it manually for each event type
it's done before generating an event so it's always done.

Resolves: #1914
