+++ TYPO3.Neos/Classes/TYPO3/Neos/EventLog/Integrations/EntityIntegrationService.php
@@ -18,7 +18,6 @@ use TYPO3\Eel\Exception;
 use TYPO3\Eel\Utility;
 use TYPO3\Flow\Annotations as Flow;
 use TYPO3\Flow\Log\SystemLoggerInterface;
-use TYPO3\Neos\EventLog\Domain\Service\EventEmittingService;
 
 /**
  * Monitors entity changes
@@ -35,12 +34,6 @@ class EntityIntegrationService extends AbstractIntegrationService
      * @var ObjectManager
      */
     protected $entityManager;
-
-    /**
-     * @Flow\Inject
-     * @var EventEmittingService
-     */
-    protected $eventEmittingService;
 
     /**
      * @Flow\Inject(lazy=FALSE)
@@ -97,7 +90,6 @@ class EntityIntegrationService extends AbstractIntegrationService
                 $entityMonitoringConfiguration = $this->monitorEntitiesSetting[$className];
 
                 if (isset($entityMonitoringConfiguration['events']['created'])) {
-                    $this->initializeAccountIdentifier();
                     $data = array();
                     foreach ($entityMonitoringConfiguration['data'] as $key => $eelExpression) {
                         $data[$key] = Utility::evaluateEelExpression($eelExpression, $this->eelEvaluator, array('entity' => $entity));
@@ -115,7 +107,6 @@ class EntityIntegrationService extends AbstractIntegrationService
                 $entityMonitoringConfiguration = $this->monitorEntitiesSetting[$className];
 
                 if (isset($entityMonitoringConfiguration['events']['deleted'])) {
-                    $this->initializeAccountIdentifier();
                     $data = array();
                     foreach ($entityMonitoringConfiguration['data'] as $key => $eelExpression) {
                         $data[$key] = Utility::evaluateEelExpression($eelExpression, $this->eelEvaluator, array('entity' => $entity));
